+++ src/subcommands/mod.rs
@@ -1,5 +1,4 @@
 pub mod backlink;
-pub mod graph;
 pub mod link;
 pub mod mv;
 pub mod new;