+++ tests/subcommands/mod.rs
@@ -1,5 +1,4 @@
 pub mod test_backlink;
-pub mod test_graph;
 pub mod test_link;
 pub mod test_mv;
 pub mod test_new;